Fixed minor warnings.

+++ code/project1_rewrite/mymalloc.h
@@ -5,10 +5,10 @@
 // STRUCTURES
 typedef struct header
 {
-    size_t size;
+    int size;
     int is_occupied;
 } HEADER;
 
 // FUNCTIONS
-void *my_malloc(size_t size, int line, char *file);
+void *my_malloc(int size, int line, char *file);
 void my_free(void *ptr, int line, char *file);
